+++ pkg/sass-parser/lib/src/statement/generic-at-rule.test.ts
@@ -49,32 +49,14 @@ describe('a generic @-rule', () => {
         () => sass.parse('@foo').nodes[0] as GenericAtRule,
       );
 
-      describe('constructed manually', () => {
-        describeNode(
-          'with a name interpolation',
-          () =>
-            new GenericAtRule({
-              nameInterpolation: new Interpolation({nodes: ['foo']}),
-            }),
-        );
+      describeNode(
+        'constructed manually',
+        () => new GenericAtRule({nameInterpolation: 'foo'}),
+      );
 
-        describeNode(
-          'with a name string',
-          () => new GenericAtRule({name: 'foo'}),
-        );
-      });
-
-      describe('constructed from ChildProps', () => {
-        describeNode('with a name interpolation', () =>
-          utils.fromChildProps({
-            nameInterpolation: new Interpolation({nodes: ['foo']}),
-          }),
-        );
-
-        describeNode('with a name string', () =>
-          utils.fromChildProps({name: 'foo'}),
-        );
-      });
+      describeNode('constructed from ChildProps', () =>
+        utils.fromChildProps({nameInterpolation: 'foo'}),
+      );
     });
 
     describe('with params', () => {
@@ -111,34 +93,14 @@ describe('a generic @-rule', () => {
         () => sass.parse('@foo bar').nodes[0] as GenericAtRule,
       );
 
-      describe('constructed manually', () => {
-        describeNode(
-          'with an interpolation',
-          () =>
-            new GenericAtRule({
-              name: 'foo',
-              paramsInterpolation: new Interpolation({nodes: ['bar']}),
-            }),
-        );
+      describeNode(
+        'constructed manually',
+        () => new GenericAtRule({name: 'foo', paramsInterpolation: 'bar'}),
+      );
 
-        describeNode(
-          'with a param string',
-          () => new GenericAtRule({name: 'foo', params: 'bar'}),
-        );
-      });
-
-      describe('constructed from ChildProps', () => {
-        describeNode('with an interpolation', () =>
-          utils.fromChildProps({
-            name: 'foo',
-            paramsInterpolation: new Interpolation({nodes: ['bar']}),
-          }),
-        );
-
-        describeNode('with a param string', () =>
-          utils.fromChildProps({name: 'foo', params: 'bar'}),
-        );
-      });
+      describeNode('constructed from ChildProps', () =>
+        utils.fromChildProps({name: 'foo', paramsInterpolation: 'bar'}),
+      );
     });
   });
 
@@ -225,7 +187,7 @@ describe('a generic @-rule', () => {
           () =>
             new GenericAtRule({
               name: 'foo',
-              paramsInterpolation: new Interpolation({nodes: ['bar ']}),
+              paramsInterpolation: 'bar ',
               nodes: [],
             }),
         );
@@ -243,7 +205,7 @@ describe('a generic @-rule', () => {
         describeNode('with an interpolation', () =>
           utils.fromChildProps({
             name: 'foo',
-            paramsInterpolation: new Interpolation({nodes: ['bar ']}),
+            paramsInterpolation: 'bar ',
             nodes: [],
           }),
         );